backend/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary

pydantic[email]

python-dotenv

passlib[bcrypt]

python-jose[cryptography]
